+++ src/main/java/madsciencemod/cards/Lightbulb.java
@@ -15,20 +15,21 @@ public class Lightbulb extends AbstractMadScienceCard {
     public static final String DESCRIPTION = cardStrings.DESCRIPTION;
     public static final String UPGRADE_DESCRIPTION = cardStrings.UPGRADE_DESCRIPTION;
     private static final int COST = 0;
-    private static final int FUEL_COST = 2;
-    private static final int UPGRADE_FUEL_COST = -1;
+    private static final int FUEL_COST = 1;
     private static final int ENERGY_AMT = 2;
+    private static final int UPGRADE_ENERGY = 1;
     private static final CardType TYPE = CardType.SKILL;
     private static final CardRarity RARITY = CardRarity.UNCOMMON;
 
     public Lightbulb() {
         super(ID, NAME, COST, DESCRIPTION, TYPE, RARITY, CardTarget.SELF);
-        this.fuelCost = this.magicNumber = this.baseMagicNumber = FUEL_COST;
+        this.fuelCost = FUEL_COST;
+        this.magicNumber = this.baseMagicNumber = ENERGY_AMT;
     }
 
     @Override
     public void use(AbstractPlayer p, AbstractMonster m) {
-        AbstractDungeon.actionManager.addToBottom(new GainEnergyAction(ENERGY_AMT));
+        AbstractDungeon.actionManager.addToBottom(new GainEnergyAction(this.magicNumber));
     }
 
     @Override
@@ -40,8 +41,9 @@ public class Lightbulb extends AbstractMadScienceCard {
     public void upgrade() {
         if (!this.upgraded) {
             upgradeName();
-            this.upgradeMagicNumber(UPGRADE_FUEL_COST);
-            this.fuelCost = this.magicNumber;
+            this.upgradeMagicNumber(UPGRADE_ENERGY);
+            this.rawDescription = UPGRADE_DESCRIPTION;
+            this.initializeDescription();
         }
     }
 }